Added I/O restrictions, Multi user chat handling and modularisation.
